cleanup in various files
