+++ packages/vscode/src/fragmentcommands.ts
@@ -11,119 +11,11 @@ import { groupBy, logInfo, logVerbose } from "../../core/src/util"
 import { resolveCli } from "./config"
 import { YAMLStringify } from "../../core/src/yaml"
 import { dotGenaiscriptPath } from "../../core/src/workdir"
-
-type TemplateQuickPickItem = {
-    template?: PromptScript
-    action?: "create"
-} & vscode.QuickPickItem
-
-async function showPromptParametersQuickPicks(
-    script: PromptScript
-): Promise<PromptParameters> {
-    const parameters: PromptParameters = {}
-    if (!script?.parameters) return {}
-
-    for (const param in script.parameters || {}) {
-        const schema = promptParameterTypeToJSONSchema(script.parameters[param])
-        switch (schema.type) {
-            case "string": {
-                let value: string
-                const enums = schema.enum
-                const uiSuggestions = schema.uiSuggestions
-                if (enums?.length) {
-                    const res = await vscode.window.showQuickPick(
-                        enums.map((e) => ({
-                            label: e,
-                            description: e,
-                        })),
-                        {
-                            title: `Choose ${schema.title || param} ${schema.description || ""}`,
-                            placeHolder: schema.default,
-                            canPickMany: false,
-                        }
-                    )
-                    value = res?.label
-                } else if (uiSuggestions) {
-                    const custom = "Enter a custom value"
-                    const res = await vscode.window.showQuickPick(
-                        [
-                            ...uiSuggestions.map((e) => ({
-                                label: e,
-                                description: e,
-                            })),
-                            {
-                                label: custom,
-                                description:
-                                    "Enter a custom value not in the suggestions.",
-                            },
-                        ],
-                        {
-                            title: `Choose ${schema.title || param} ${schema.description || ""}`,
-                            placeHolder: schema.default,
-                            canPickMany: false,
-                        }
-                    )
-                    value = res?.label
-                    if (value === custom) {
-                        value = await vscode.window.showInputBox({
-                            title: `Enter value for ${schema.title || param}`,
-                            value: schema.default,
-                            prompt: schema.description,
-                        })
-                    }
-                    if (value === undefined) return undefined
-                } else {
-                    value = await vscode.window.showInputBox({
-                        title: `Enter value for ${schema.title || param}`,
-                        value: schema.default,
-                        prompt: schema.description,
-                    })
-                }
-                if (value === undefined) return undefined
-                parameters[param] = value
-                break
-            }
-            case "boolean": {
-                const value = await vscode.window.showQuickPick(
-                    [{ label: "yes" }, { label: "no" }],
-                    {
-                        title: `Choose ${schema.title || param} ${schema.description || ""}`,
-                        canPickMany: false,
-                    }
-                )
-                if (value === undefined) return undefined
-                parameters[param] = value.label === "yes"
-                break
-            }
-            case "integer":
-            case "number": {
-                const parse = schema.type === "integer" ? parseInt : parseFloat
-                const value = await vscode.window.showInputBox({
-                    title: `Enter ${schema.type} value for ${schema.title || param}`,
-                    value: schema.default?.toString(),
-                    prompt: schema.description,
-                    validateInput: (v) => {
-                        v = v.trim()
-                        const x = parse(v)
-                        const msg =
-                            isNaN(x) || String(x) !== v
-                                ? `Enter a valid, finite ${schema.type}`
-                                : null
-                        return msg
-                    },
-                })
-                if (value === undefined) return undefined
-                parameters[param] = parse(value)
-                break
-            }
-            default:
-                throw new NotSupportedError(
-                    "Unsupported parameter type " + schema.type
-                )
-        }
-    }
-    return parameters
-}
+import {
+    TemplateQuickPickItem,
+    showPromptParametersQuickPicks,
+} from "./parameterquickpick"
+import { templatesToQuickPickItems } from "./scriptquickpick"
 
 export function activateFragmentCommands(state: ExtensionState) {
     const { context, host } = state
@@ -303,49 +195,4 @@ export function activateFragmentCommands(state: ExtensionState) {
         registerCommand("genaiscript.fragment.prompt", fragmentPrompt),
         registerCommand("genaiscript.fragment.debug", fragmentDebug)
     )
-}
-
-export function templatesToQuickPickItems(
-    templates: globalThis.PromptScript[],
-    options?: { create?: boolean }
-): TemplateQuickPickItem[] {
-    const { create } = options || {}
-    const cats = groupBy(templates, templateGroup)
-    const items: vscode.QuickPickItem[] = []
-    for (const cat in cats) {
-        items.push(<vscode.QuickPickItem>{
-            label: cat,
-            kind: vscode.QuickPickItemKind.Separator,
-        })
-        items.push(
-            ...cats[cat].map(
-                (template) =>
-                    <TemplateQuickPickItem>{
-                        label:
-                            template.title ??
-                            (template.filename &&
-                                vscode.workspace.asRelativePath(
-                                    template.filename
-                                )) ??
-                            template.id,
-                        description: `${template.id} ${
-                            template.description || ""
-                        }`,
-                        template,
-                    }
-            )
-        )
-    }
-    if (create) {
-        items.push(<vscode.QuickPickItem>{
-            label: "",
-            kind: vscode.QuickPickItemKind.Separator,
-        })
-        items.push(<TemplateQuickPickItem>{
-            label: "Create a new GenAIScript script...",
-            description: "Create a new script script in the current workspace.",
-            action: "create",
-        })
-    }
-    return items
 }
